Renamed and Relocated RequestModifierExtensions to RequestModifier.js Relocated VirtualBuffer.js to streaming root.
