+++ src/streaming/MediaPlayer.js
@@ -38,8 +38,8 @@ import ErrorHandler from './ErrorHandler.js';
 import Capabilities from './utils/Capabilities.js';
 import TextTrackExtensions from './extensions/TextTrackExtensions.js';
 import SourceBufferExtensions from './extensions/SourceBufferExtensions.js';
-import VirtualBuffer from './utils/VirtualBuffer.js';
-import RequestModifierExtensions from './extensions/RequestModifierExtensions.js';
+import VirtualBuffer from './VirtualBuffer.js';
+import RequestModifier from './utils/RequestModifier.js';
 import TextSourceBuffer from './TextSourceBuffer.js';
 import URIQueryAndFragmentModel from './models/URIQueryAndFragmentModel.js';
 import ManifestModel from './models/ManifestModel.js';
@@ -1586,7 +1586,7 @@ function MediaPlayer() {
             errHandler: errHandler,
             parser: createManifestParser(),
             metricsModel: metricsModel,
-            requestModifierExt: RequestModifierExtensions(context).getInstance()
+            requestModifier: RequestModifier(context).getInstance()
         });
     }
